pyproject.toml
[tool.poetry]
name = "ses-statistics-web"
version = "0.1.0"
description = "Web-app for vizualize processed by 'SES' data."
authors = ["Kubanychbek uulu Kairat <kubanychbekukairat@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pandas = "^2.2.2"
numpy = "^2.1.0"
requests = "^2.32.3"
streamlit = "^1.38.0"
fastapi = "^0.112.2"
plotly = "^5.23.0"
streamlit-aggrid = "^1.0.5"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
